feat: Add useAuthState hook
